requirements.txt
alembic==1.0.11
amqp==2.5.0
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Celery==2.4.3
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Spyne==0.3.1
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
lxml==4.3.4
Mako==1.0.14
MarkupSafe==1.1.1
mock==2.0.0
more-itertools==7.2.0
pbr==5.4.1
pluggy==0.6.0
py==1.8.0
PyMySQL==0.7.11
pytest==3.6.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
requests==2.22.0
sec-wall==1.2
six==1.12.0
spyne==2.12.16
SQLAlchemy==1.3.6
suds-jurko==0.6
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
